move various modules to core
