+++ src/client.rs
@@ -15,29 +15,28 @@
  * limitations under the License.
  */
 
-use crate::arena;
-use crate::buffer::TmpBuffer;
 use crate::can_move_mio_sockets_between_threads;
-use crate::channel;
 use crate::connection::{
     client_req_connection, client_stream_connection, ConnectionPool, StreamSharedData,
 };
+use crate::core::arena;
+use crate::core::buffer::TmpBuffer;
+use crate::core::channel;
+use crate::core::event;
+use crate::core::executor::{Executor, Spawner};
+use crate::core::list;
+use crate::core::reactor::Reactor;
+use crate::core::tnetstring;
+use crate::core::zmq::{MultipartHeader, SpecInfo};
 use crate::counter::Counter;
-use crate::event;
-use crate::executor::{Executor, Spawner};
 use crate::future::{
     event_wait, select_2, select_5, select_6, select_option, yield_to_local_events,
     AsyncLocalReceiver, AsyncLocalSender, AsyncReceiver, CancellationSender, CancellationToken,
     Select2, Select5, Select6, Timeout,
 };
-use crate::list;
-use crate::pin;
-use crate::reactor::Reactor;
 use crate::resolver::Resolver;
-use crate::tnetstring;
 use crate::zhttppacket;
 use crate::zhttpsocket::{self, SessionKey, FROM_MAX, REQ_ID_MAX};
-use crate::zmq::{MultipartHeader, SpecInfo};
 use arrayvec::ArrayVec;
 use ipnet::IpNet;
 use log::{debug, error, info, warn};
@@ -49,6 +48,7 @@ use std::collections::{HashMap, VecDeque};
 use std::convert::TryFrom;
 use std::io::{self, Write};
 use std::mem;
+use std::pin::pin;
 use std::rc::Rc;
 use std::str;
 use std::sync::{mpsc, Arc};
