+++ src/core/list.rs
