+++ src/core/shuffle.rs
