+++ src/lib.rs
@@ -21,41 +21,26 @@
  * $FANOUT_END_LICENSE$
  */
 
-pub mod arena;
-pub mod buffer;
-pub mod channel;
 pub mod client;
-pub mod config;
 pub mod connection;
 pub mod connmgr;
 pub mod core;
 pub mod counter;
-pub mod event;
-pub mod executor;
 pub mod ffi;
 pub mod future;
 pub mod jwt;
-pub mod list;
 pub mod listener;
-pub mod log;
-pub mod net;
 pub mod pool;
 pub mod publish_cli;
-pub mod reactor;
 pub mod resolver;
 pub mod runner;
 pub mod server;
 pub mod service;
-pub mod shuffle;
-pub mod timer;
 pub mod tls;
-pub mod tnetstring;
 pub mod track;
-pub mod waker;
 pub mod websocket;
 pub mod zhttppacket;
 pub mod zhttpsocket;
-pub mod zmq;
 
 pub use std::pin::pin;
